Move static methods from FunctionRegistry to FunctionUtils
